Overhauled the interpreter
The character-by-character iterator was replaced with a pattern that
matches an MML note or command. Since that pattern relies on there being
another note or command after the current one, there is a special case
for the last note. If I can figure out a pattern that will work there, I
will change it.

This change allowed the removal of a lot of "string.sub"s when looking
for arguments (eg. a note being sharp, note length, t[bpm], etc.) and
generally made the code neater.